+++ libs/lit-vanilla/src/lib/components/textinput.element.ts
@@ -2,7 +2,7 @@ import * as Core from '@golemui/core';
 import * as Lit from '@golemui/lit';
 import { TextinputProps } from '@golemui/shared-vanilla';
 import { consume, provide } from '@lit/context';
-import { LitElement, html, nothing } from 'lit';
+import { html, LitElement, nothing } from 'lit';
 import { customElement, property } from 'lit/decorators.js';
 import { classMap } from 'lit/directives/class-map.js';
 import { Subscription } from 'rxjs';
@@ -66,6 +66,11 @@ export class TextinputElement extends LitElement implements Core.WithField {
       icon = html``;
     }
 
+    const showErrors =
+      this.adapter.templateData.touched &&
+      this.adapter.templateData.errors &&
+      this.adapter.templateData.errors.length > 0;
+
     return html`
       <label for=${this.field.uid}>
         ${this.adapter.templateData.label +
@@ -84,15 +89,18 @@ export class TextinputElement extends LitElement implements Core.WithField {
           placeholder=${this.adapter.templateData.placeholder || nothing}
           @input="${() => this.valueChanged(event)}"
           @blur="${() => this.adapter.onBlur()}"
+          aria-invalid=${showErrors || nothing}
           aria-required=${this.adapter.templateData.validator?.required || nothing}
           aria-describedby=${this.adapter.templateData.hint ? `${this.field.uid}_hint` : nothing}
         />
         ${icon}
       </div>
 
-      ${this.adapter.templateData.errors && this.adapter.templateData.errors.length > 0
-        ? html`<ul>
-            ${this.adapter.templateData.errors.map((error) => html`<li>${error}</li>`)}
+      ${showErrors
+        ? html`<ul class="gui-validator">
+            ${this.adapter.templateData.errors?.map(
+              (error) => html`<li class="gui-validator__error">${error}</li>`,
+            )}
           </ul>`
         : ''}
     `;