+++ libs/lit/src/lib/adapters/control.field-adapter.ts
@@ -44,6 +44,11 @@ export class ControlFieldAdapter<
         });
       });
 
+    // Listen to the touchedControls stream for this control
+    this.context.store.state$
+      .pipe(takeUntil(this.destroy$), Core.touchedControlsByPath$(field.path))
+      .subscribe((touched) => this.setTemplateData({ touched }));
+
     // Listen to the fieldFlags stream (`disabled`, `required` and `readonly` flags)
     this.context.store.state$
       .pipe(takeUntil(this.destroy$), Core.fieldFlagsByUid$(field.uid))
@@ -83,7 +88,8 @@ export class ControlFieldAdapter<
 
   onBlur() {
     this.context.store.dispatch({
-      type: 'TOUCHED',
+      type: 'ATTEMPT_VALIDATION',
+      payload: { reason: 'blur', path: this.field.path },
     });
   }